refactor(location): refactor project service
move project service to /shared/services
