+++ src/app/project/project.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
-import { Project } from './project';
-import { ProjectService } from './project.service';
+import { Project } from '../shared/services/project';
+import { ProjectService } from '../shared/services/project.service';
 
 
 @Component({
