+++ src/app/project/project.service.spec.ts
@@ -1,7 +1,7 @@
 /* tslint:disable:no-unused-variable */
 
 import { TestBed, async, inject } from '@angular/core/testing';
-import { ProjectService } from './project.service';
+import { ProjectService } from '../shared/services/project.service';
 
 describe('Service: Project', () => {
   beforeEach(() => {
